Bring the breadcrumb up further and mapStateToProps up another level as well.
